chatbot/requirements.txt
flask==2.3.2
flask-cors==3.0.10
openai==1.0.0
pinecone-client==5.0.1
sentence-transformers==2.2.2
langdetect==1.0.9
python-dotenv==1.0.0
google-cloud-translate==3.11.0
nltk==3.8.1
uvicorn==0.22.0
gunicorn==20.1.0
requests==2.31.0
sqlalchemy==2.0.19
pymysql==1.0.3
flask-sqlalchemy==3.0.3
flask-login==0.6.3
pydantic==2.8.0
